requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email_validator==2.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.0
pytest-cov==4.1.0
